General code cleanup and speedup
- Reduce unnecessary class loader creation in deploy artifact.
  - Speed up artifact deployment
- Move AppJarHelper and PluginJarHelper to test source
  - They should strictly be used in unit-test only
- Cleanup logback xmls
  - Remove the no longer in used hacky CAppender, LogWriter and LogCollector
- Reduce AuthorizationService log level on cache usage
  - Caching is enabled by default (and is always needed), no point
    in logging a warn when authorization is turned off.
- Fix race conditions in authorization cache unit-tests
