+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/services/ApplicationLifecycleServiceTest.java
@@ -19,10 +19,10 @@ package co.cask.cdap.internal.app.services;
 import co.cask.cdap.MissingMapReduceWorkflowApp;
 import co.cask.cdap.common.ArtifactNotFoundException;
 import co.cask.cdap.common.io.Locations;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.internal.app.deploy.ProgramTerminator;
 import co.cask.cdap.internal.app.runtime.artifact.ArtifactRepository;
 import co.cask.cdap.internal.app.services.http.AppFabricTestBase;
-import co.cask.cdap.internal.test.AppJarHelper;
 import co.cask.cdap.proto.Id;
 import com.google.common.io.Files;
 import org.apache.twill.filesystem.Location;
