+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/services/ProgramLifecycleServiceAuthorizationTest.java
@@ -19,8 +19,8 @@ package co.cask.cdap.internal.app.services;
 import co.cask.cdap.AllProgramsApp;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.internal.AppFabricTestHelper;
-import co.cask.cdap.internal.test.AppJarHelper;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.ProgramType;
 import co.cask.cdap.proto.id.InstanceId;
