+++ cdap-app-fabric/src/test/resources/logback-test.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <!--
-  Copyright © 2014 Cask Data, Inc.
+  Copyright © 2014-2016 Cask Data, Inc.
 
   Licensed under the Apache License, Version 2.0 (the "License"); you may not
   use this file except in compliance with the License. You may obtain a copy of
@@ -34,17 +34,9 @@
         </encoder>
     </appender>
 
-    <appender name="CLOG" class="co.cask.cdap.common.logging.logback.CAppender">              
-        <layout>                      
-            <!-- Note: we don't insert new line in the pattern as this is done on higher level (in old back-end) -->
-            <pattern>%d{ISO8601} - %-5p [%t:%C{1}@%L] - %m</pattern>       
-        </layout>
-    </appender>
-
     <logger name="co.cask.cdap" level="INFO" />
 
     <root level="WARN">
-        <appender-ref ref="CLOG"/>
         <appender-ref ref="STDOUT"/>
     </root>
