+++ cdap-cli-tests/src/test/java/co/cask/cdap/cli/CLIMainTest.java
@@ -36,8 +36,8 @@ import co.cask.cdap.common.DatasetTypeNotFoundException;
 import co.cask.cdap.common.ProgramNotFoundException;
 import co.cask.cdap.common.UnauthenticatedException;
 import co.cask.cdap.common.io.Locations;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.explore.client.ExploreExecutionResult;
-import co.cask.cdap.internal.test.AppJarHelper;
 import co.cask.cdap.proto.DatasetTypeMeta;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.NamespaceMeta;
