+++ cdap-client-tests/src/test/java/co/cask/cdap/client/ArtifactClientTestRun.java
@@ -29,9 +29,9 @@ import co.cask.cdap.client.util.RESTClient;
 import co.cask.cdap.common.ArtifactNotFoundException;
 import co.cask.cdap.common.BadRequestException;
 import co.cask.cdap.common.NotFoundException;
+import co.cask.cdap.common.test.AppJarHelper;
+import co.cask.cdap.common.test.PluginJarHelper;
 import co.cask.cdap.internal.io.ReflectionSchemaGenerator;
-import co.cask.cdap.internal.test.AppJarHelper;
-import co.cask.cdap.internal.test.PluginJarHelper;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.artifact.ApplicationClass;
 import co.cask.cdap.proto.artifact.ApplicationClassInfo;
