+++ cdap-client-tests/src/test/resources/logback-test.xml
@@ -35,17 +35,9 @@
     </encoder>
   </appender>
 
-  <appender name="CLOG" class="co.cask.cdap.common.logging.logback.CAppender">              
-    <layout>                      
-      <!-- Note: we don't insert new line in the pattern as this is done on higher level (in old back-end) -->
-      <pattern>%d{ISO8601} - %-5p [%t:%C{1}@%L] - %m</pattern>       
-    </layout>
-  </appender>
-
   <logger name="co.cask.cdap" level="DEBUG" />
 
   <root level="DEBUG">
-    <appender-ref ref="CLOG"/>
     <appender-ref ref="STDOUT"/>
   </root>
