+++ cdap-common/src/test/java/co/cask/cdap/common/lang/ClassLoaderTest.java
@@ -23,8 +23,8 @@ import co.cask.cdap.api.app.ApplicationConfigurer;
 import co.cask.cdap.api.common.Bytes;
 import co.cask.cdap.common.io.Locations;
 import co.cask.cdap.common.lang.jar.BundleJarUtil;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.internal.io.SchemaGenerator;
-import co.cask.cdap.internal.test.AppJarHelper;
 import com.google.common.base.Predicate;
 import com.google.common.base.Predicates;
 import com.google.common.collect.ImmutableList;
