+++ cdap-explore/src/test/java/co/cask/cdap/explore/service/ExploreServiceUtilsTest.java
@@ -37,9 +37,7 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.lang.reflect.Method;
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.net.URL;
-import java.nio.file.Paths;
 import java.security.CodeSource;
 import java.util.jar.JarEntry;
 import java.util.jar.JarFile;
