+++ cdap-gateway/src/test/java/co/cask/cdap/gateway/GatewayFastTestsSuite.java
@@ -19,12 +19,12 @@ package co.cask.cdap.gateway;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.common.io.Locations;
 import co.cask.cdap.common.lang.jar.BundleJarUtil;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.common.utils.DirUtils;
 import co.cask.cdap.gateway.handlers.PingHandlerTestRun;
 import co.cask.cdap.gateway.handlers.RuntimeArgumentTestRun;
 import co.cask.cdap.gateway.handlers.hooks.MetricsReporterHookTestRun;
 import co.cask.cdap.gateway.run.StreamWriterTestRun;
-import co.cask.cdap.internal.test.AppJarHelper;
 import com.google.common.base.Charsets;
 import com.google.common.collect.ObjectArrays;
 import com.google.common.io.Files;
