+++ cdap-security/src/test/java/co/cask/cdap/security/authorization/DefaultAuthorizationEnforcementServiceTest.java
@@ -19,8 +19,8 @@ package co.cask.cdap.security.authorization;
 import co.cask.cdap.api.Predicate;
 import co.cask.cdap.common.conf.CConfiguration;
 import co.cask.cdap.common.conf.Constants;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.common.utils.Tasks;
-import co.cask.cdap.internal.test.AppJarHelper;
 import co.cask.cdap.proto.id.DatasetId;
 import co.cask.cdap.proto.id.EntityId;
 import co.cask.cdap.proto.id.InstanceId;
