+++ cdap-unit-test/src/main/java/co/cask/cdap/test/UnitTestManager.java
@@ -34,13 +34,13 @@ import co.cask.cdap.common.discovery.StickyEndpointStrategy;
 import co.cask.cdap.common.io.Locations;
 import co.cask.cdap.common.lang.ProgramResources;
 import co.cask.cdap.common.namespace.NamespaceAdmin;
+import co.cask.cdap.common.test.AppJarHelper;
+import co.cask.cdap.common.test.PluginJarHelper;
 import co.cask.cdap.data2.dataset2.DatasetFramework;
 import co.cask.cdap.explore.jdbc.ExploreDriver;
 import co.cask.cdap.internal.AppFabricClient;
 import co.cask.cdap.internal.app.runtime.artifact.ArtifactRepository;
 import co.cask.cdap.internal.app.runtime.artifact.Artifacts;
-import co.cask.cdap.internal.test.AppJarHelper;
-import co.cask.cdap.internal.test.PluginJarHelper;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.NamespaceMeta;
 import co.cask.cdap.proto.artifact.AppRequest;
