+++ cdap-unit-test/src/test/java/co/cask/cdap/test/base/TestFrameworkTestBase.java
@@ -17,7 +17,7 @@
 package co.cask.cdap.test.base;
 
 import co.cask.cdap.api.app.Application;
-import co.cask.cdap.internal.test.AppJarHelper;
+import co.cask.cdap.common.test.AppJarHelper;
 import co.cask.cdap.proto.artifact.AppRequest;
 import co.cask.cdap.proto.artifact.ArtifactSummary;
 import co.cask.cdap.proto.id.ArtifactId;
